ai/pyproject.toml
[project]
name = "mizhou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
]